+++ src/components/Spinner.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import '../styles/Spinner.css'
 
 const Spinner: React.FC = () => {
+  console.log("Spinner Component");
   return <div className="spinner"></div>;
 };