+++ src/hooks/useQuery.ts
@@ -1,0 +1,5 @@
+const useQuery = ():URLSearchParams => {
+    return new URLSearchParams(window.location.search); 
+  };
+
+export default useQuery;