+++ src/types/List.d.ts
@@ -1,0 +1,7 @@
+
+export default interface ListProps {
+    items: string[],
+    listClassName : string, 
+    listItemClassName: string,
+    selectedIndex: number
+  }